moved ColorLegend into Interface namespace
