--- 3DAmsterdam/Assets/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6b3b45e06917f3e43b214de74d4019da
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
